chore(bootstrap): bump the seed

+++ bootstrap/seed/darwin-x86_64/bootstrap.std_io.c
@@ -74,7 +74,7 @@ static void std_io_write(int32_t l0, vader_string_t l1) {
 
 static int32_t std_io_write_bytes(int32_t l0, void* l1, size_t l2) {
     int32_t l3;
-    int32_t t0;
+    ptrdiff_t t0;
     void** gc_raw_roots[1] = { &l1 };
     vader_gc_frame_t gc_frame = { vader_gc_top, 0u, 1u, NULL, gc_raw_roots, 0u, NULL };
     vader_gc_top = &gc_frame;
@@ -84,7 +84,7 @@ static int32_t std_io_write_bytes(int32_t l0, void* l1, size_t l2) {
         l3 = INT32_C(2);
     }
     t0 = vader_host_system_posix_sys_write(l3, l1, l2);
-    { int32_t __vret = t0; vader_gc_top = gc_frame.prev; return __vret; }
+    { int32_t __vret = ((int32_t) (ptrdiff_t) t0); vader_gc_top = gc_frame.prev; return __vret; }
     vader_gc_top = gc_frame.prev;
 }
